requirements.txt
pandas
scikit-learn
matplotlib
joblib
yfinance
python-telegram-bot==13.15
